+++ example/oglplus/027_depth_of_field.cpp
@@ -203,7 +203,7 @@ public:
 			attr.Enable();
 		}
 
-		Uniform<Vec3f>(main_prog, "LightPos").Set(30.0, 50.0, 20.0);
+		Uniform<Vec3f>(main_prog, "LightPos").Set(30.0f, 50.0f, 20.0f);
 
 		dof_vs.Source(
 			"#version 330\n"
@@ -412,7 +412,7 @@ public:
 		dof_prog.Use();
 		gl.Bind(screen);
 
-		focus_depth.Set(0.6 + SineWave(time / 9.0)*0.3);
+		focus_depth.Set(GLfloat(0.6 + SineWave(time / 9.0)*0.3));
 
 		gl.Enable(Capability::Blend);
 		gl.DrawArrays(PrimitiveType::TriangleStrip, 0, 4);
